sound/requirements.txt
PyQt6
markdown
edge-tts
pygame
pyyaml
nest_asyncio
mutagen
